2023-18
TIL: Shoelace formula can be used to find area of a shape given a list
of vertices in CW or CCW order. Since this question deals with a grid
however, all points are in the middle of the grid. Using Pick's theorem
can then get us the number of inner full tiles, and when added to the
length of the perimeter, will give the total area.